+++ src/lib/collapseAdjacentRules.js
