+++ src/lib/setupContextUtils.js
@@ -4,14 +4,14 @@ import postcss from 'postcss'
 import dlv from 'dlv'
 import selectorParser from 'postcss-selector-parser'
 
-import transformThemeValue from '../../util/transformThemeValue'
-import parseObjectStyles from '../../util/parseObjectStyles'
-import prefixSelector from '../../util/prefixSelector'
-import isPlainObject from '../../util/isPlainObject'
-import escapeClassName from '../../util/escapeClassName'
-import nameClass from '../../util/nameClass'
-import { coerceValue } from '../../util/pluginUtils'
-import bigSign from '../../util/bigSign'
+import transformThemeValue from '../util/transformThemeValue'
+import parseObjectStyles from '../util/parseObjectStyles'
+import prefixSelector from '../util/prefixSelector'
+import isPlainObject from '../util/isPlainObject'
+import escapeClassName from '../util/escapeClassName'
+import nameClass from '../util/nameClass'
+import { coerceValue } from '../util/pluginUtils'
+import bigSign from '../util/bigSign'
 import corePlugins from '../corePlugins'
 import * as sharedState from './sharedState'
 import { env } from './sharedState'
