+++ src/lib/sharedState.js
