+++ src/processTailwindFeatures.js
@@ -1,8 +1,8 @@
 import normalizeTailwindDirectives from './lib/normalizeTailwindDirectives'
 import expandTailwindAtRules from './lib/expandTailwindAtRules'
 import expandApplyAtRules from './lib/expandApplyAtRules'
-import evaluateTailwindFunctions from '../lib/evaluateTailwindFunctions'
-import substituteScreenAtRules from '../lib/substituteScreenAtRules'
+import evaluateTailwindFunctions from './lib/evaluateTailwindFunctions'
+import substituteScreenAtRules from './lib/substituteScreenAtRules'
 import resolveDefaultsAtRules from './lib/resolveDefaultsAtRules'
 import collapseAdjacentRules from './lib/collapseAdjacentRules'
 import { createContext } from './lib/setupContextUtils'
