+++ tests/jit/custom-transformers.test.js
@@ -3,7 +3,7 @@ import path from 'path'
 
 function run(input, config = {}) {
   jest.resetModules()
-  const tailwind = require('../../src/jit/index.js').default
+  const tailwind = require('../../src')
   return postcss(tailwind(config)).process(input, {
     from: path.resolve(__filename),
   })
